refresh: Generate archived Dockerfiles
We need this so that we can perform Quay builds without going
through a build trigger.

Once we start generating Dockerfiles this way, each refresh will
result in a bunch of binary files being modified, which is not
very nice to humans following along at home.

Thankfully, there is a way the issue can be mitigated: just
configure git with

  # ~/.config/git/attributes
  buildenv-*.zip diff=dockerfilezip

  # ~/.config/git/config
  [diff "dockerfilezip"]
  binary = true
  textconv = unzip -q -c -a

after which 'git diff' and friends will automatically show the
contents of the archive.

Note that the configuration above only works correctly if there
is a single file in the archive, but for us that's the case by
definition so it's not really a problem.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Fabiano Fidêncio <fidencio@redhat.com>